set user IdentificationNumber as long

+++ StandUsers.Domain/User/Dtos/UserDto.cs
@@ -15,5 +15,5 @@ public class UserDto
     public string Direction { get; set; }
 
     [Required]
-    public int IdentificationNumber { get; set; }
+    public long IdentificationNumber { get; set; }
 }
