+++ StandUsers.Domain/User/Dtos/UserOwnedDto.cs
@@ -4,5 +4,5 @@ public class UserOwnedDto
 {
     public Guid Id { get; set; }
     public string Email { get; set; }
-    public int IdentificationNumber { get; set; }
+    public long IdentificationNumber { get; set; }
 }
